// src/components/library/button/primary/types.ts
export interface PrimaryButtonPropsI {
    name: string;
    onClick: () => void;
    id?: string;
    width?: string;
    height?: string;
    styles?: string;
    margin?: string;
    border?: string;
    padding?: string;
    outline?: string;
    loader?: boolean;
    fontSize?: string;
    isColor?: boolean;
    className?: string;
    boxShadow?: string;
    borderRadius?: string;
    disabled?: boolean;
}